% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssTest_SNP.R
\name{AssTest_SNP}
\alias{AssTest_SNP}
\title{Single SNP Association Test}
\usage{
AssTest_SNP(gp, trait, map, weights = NULL, power = -1, mk.sel = NULL,
  K_i = TRUE)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2 and family.}

\item{trait}{\code{Numerical} vector of phenotypic trait values.}

\item{map}{Three columns \code{data.frame} with marker identifier,
chromosome, and marker position (cM or bp).}

\item{weights}{Two columns \code{data.frame} containing the marker
identifiers and the weights marker weights. For example the weights obtained
with the LDAK program (see \code{\link{LDAK_weights}}). By default, the
kinship matrix is computed unweighted or all weights are equal to 1, which
correspond to the Astle and Balding kinship matrix}

\item{power}{\code{Numerical} value specifying the value of the
parameter for marker scores standardization. The column of the marker matrix
(X.j) are multiplied by var(X.j)^(power/2). It correspond to alpha in the
formula .Default = -1.}

\item{mk.sel}{\code{Character vector} specifying a list of marker to use
for the kinship matrix computation. By default, the function use all markers
of the \code{gp}.}

\item{K_i}{\code{Logical} specifying if the kinship correction should be done
by removing the markers of the scanned chromosome. Default = TRUE.}
}
\value{
Return:

\item{results}{Four column data.frame with marker identifier, chromosome,
position and -log10(p-value).}
}
\description{
Perform a single SNP genome wide association test using a kinship matrix
to correct for the genetic background.
}
\details{
The function is a wrapper for the \code{GWAS} function from the \code{sommer}
package (Covarrubias-Pazaran, 2016). The model is fitted using the EMMA
algorithm proposed by Kang et al. (2008).

By default, the kinship matrix is computed using the method of Astle and
Balding (2009). It is possible to compute a linkage disequilibrium adjusted
kinship (LDAK) kinship matrix using the method of Speed et al. (2012) by
introducing the weights computed with the function \code{\link{LDAK_weights}}.
The model can be fitted using the kinship containing all markers or removing
the markers of the scanned chromosome (\code{K_i = TRUE}).
}
\references{
Astle, W., & Balding, D. J. (2009). Population structure and cryptic
relatedness in genetic association studies. Statistical Science, 451-471.

Covarrubias-Pazaran G. 2016. Genome assisted prediction of quantitative traits
using the R package sommer. PLoSONE 11(6):1-15.

Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.
}
\author{
Vincent Garin
}
