\name{EUNAM_gp}
\alias{EUNAM_gp}
\docType{data}
\title{
  gpData object for the reduced EUNAM example data
}
\description{
  gpData object for the reduced EUNAM example data.
}
\usage{data(EUNAM_gp)}
\format{
  \code{data.frame}
}

\examples{
  data(EUNAM_gp)
}

\seealso{
  \code{\link{EUNAM_geno}}, \code{\link{EUNAM_map}}, \code{\link{EUNAM_pheno}}
}

\keyword{datasets}
