\name{EUNAM_LD_weights}
\alias{EUNAM_LD_weights}
\docType{data}
\title{
  LDAK weights for the reduced EUNAM example data
}
\description{
  LDAK weights for the reduced EUNAM example data.
}
\usage{data(EUNAM_LD_weights)}
\format{
  \code{data.frame}
}

\examples{
  data(EUNAM_LD_weights)
}

\seealso{
  \code{\link{EUNAM_gp}}
}

\keyword{datasets}
