\name{EUNAM_map}
\alias{EUNAM_map}
\docType{data}
\title{
  Reduced map maize EU-NAM Dent population
}
\description{
  Reduced map of the maize EU nested association mapping (NAM) population
  (Bauer et al., 2013).
}
\usage{data(EUNAM_map)}
\format{
  \code{data.frame}
}
\details{

  Selection of 2182 markers from chromosomes 2 and 3 of the Maize EU-NAM
  population (Bauer et al., 2013). We used the genetic map computed by Giraud
  et al. (2014) available here http://maizegdb.org/data_center/reference?id=9024747.

  }
\source{

  http://maizegdb.org/data_center/reference?id=9024747

}
\references{
  Bauer, E., Falque, M., Walter, H., Bauland, C., Camisan, C., Campo, L., ... &
    Altmann, T. (2013). Intraspecific variation of recombination rate in maize.
  Genome biology, 14(9), R103.

  Giraud, H., Lehermeier, C., Bauer, E., Falque, M., Segura, V., Bauland, C., ...
  & Schipprack, W. (2014). Linkage disequilibrium with linkage analysis of
  multiline crosses reveals different multiallelic QTL for hybrid performance
  in the flint and dent heterotic groups of maize. Genetics, 198(4), 1717-1734.

}
\examples{
  data(EUNAM_map)
}

\seealso{
  \code{\link{EUNAM_geno}}, \code{\link{EUNAM_pheno}}
}

\keyword{datasets}
