% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_SSD_file.R
\name{generate_SSD_file}
\alias{generate_SSD_file}
\title{Kernel association test SSD files}
\usage{
generate_SSD_file(gp, map, out.dir, prefix, plink.dir, hap = 1,
  hap.unit = 1, sliding.window = FALSE, gap = 1, LD.based = FALSE,
  ld.threshold = 0.9)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2.}

\item{map}{Four columns \code{data.frame} with marker id, chromosome,
getic position in cM and physical position in bp.}

\item{out.dir}{output directory where temporary files will be stored.}

\item{prefix}{\code{Character} string. The SSD files will be saved as
prefix.SetID, prefix.SSD and prefix.SSD.info in \code{out.dir}.}

\item{plink.dir}{directory where the plink executable program is located.}

\item{hap}{Defines either a number of markers (if \code{hap.unit=1}) or a
genetic distance in cM (if \code{hap.unit=2}) that should be used for building
haplo blocks. Default = 1.}

\item{hap.unit}{Set to 1 for building haplo blocks based on number of
markers, set to 2 for building haplo blocks based on markers lying in a
specified stretch of the chromosome (in cM). Default = 1.}

\item{sliding.window}{\code{logical} value specifying if
the haplotype blocks should be construct using a sliding window. If not,
the haplotype blocks are adjacent and not overlapping
(e.g. b1 = 123, b2 = 456, etc.). Default = TRUE.}

\item{gap}{\code{Numeric} value indicating the number of markers between the
center of two consecutive haplotype blocks. Default = 1.}

\item{LD.based}{NOT AVAILABLE NOW. \code{logical} value specifying if the
haplotype blocks should be constructed based on LD. Default = FALSE.}

\item{ld.threshold}{NOT AVAILABLE NOW.}
}
\value{
Create the SSD files in the location specified in \code{out.dir}. return
also \code{map_plot} a two column data.frame with the chromosome and the
center bp position of each marker set.
}
\description{
Generate the SSD files for the kernel association test. The function
partitions the genome in adjacent or sliding window blocks based on
a fixed number of marker or cM distance.
}
\author{
Vincent Garin
}
