% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_GSM.R
\name{GWAS_GSM}
\alias{GWAS_GSM}
\title{GSM Association Test}
\usage{
GWAS_GSM(SSD_file, gp, trait = 1, kernel = "linear.weighted",
  weights.beta = c(0.5, 0.5), weights.kernel = NULL, kin.correct = TRUE,
  K_i = TRUE, weights = NULL, power = -1, n.cores = NULL,
  verbose = FALSE)
}
\arguments{
\item{SSD_file}{object of class \code{SSD_file} obtained with the function
\code{\link{generate_SSD_file}}.}

\item{gp}{\code{gpData} object with elements geno coded 0 1 2, map with
marker position in cM, phenotype and family. \strong{containing the list of
markers that will be used to compute the kinship matrix}.}

\item{trait}{\code{Numerical} or \code{character} indicator to specify which
trait of the gp object should be used. Default = 1.}

\item{kernel}{\code{Character} element indicating the type of kernel function
used to calculate the variance covariance structure using marker
set information. can e one of: "linear", "linear.weighted", "IBS",
"IBS.weighted", "quadratic" and "2wayIX". For details see details of the
function SKAT (pacakge SKAT). Default = "linear.weighted".}

\item{weights.beta}{a numeric vector of parameters for the beta weights for
the weighted kernels. Default = c(0.5, 0.5).}

\item{weights.kernel}{NOT AVAILABLE NOW. Vector of own provided weights for
for the kernel function. Default = NULL.}

\item{kin.correct}{\code{Logical} value. If \code{kin.correct = TRUE}, the
model will use a kinship matrix to correct for the genetic background.
Default = TRUE.}

\item{K_i}{\code{Logical} specifying if the kinship correction should be done
by removing the markers of the scanned chromosome. Default = TRUE.}

\item{weights}{object of class \code{LD_wgh} obtained by the function
LDAK_weights() representing a data.frame with two columns: the marker
identifier and the LD adjusted weights. These weight will be used to compute
a LDAK as defined by Speed et al. (2012) for the genetic background
adjustement. Default = NULL. By default, the kinship matrix is computed
unweighted or all weights are equal to 1, which correspond to the Astle and
Balding kinship matrix.}

\item{power}{\code{Numerical} value specifying the value of the
parameter for marker scores standardization. The column of the marker matrix
(X.j) are multiplied by var(X.j)^(power/2). It correspond to alpha in the
formula. Default = -1.}

\item{n.cores}{\code{Numeric} value indicating the number of core to be used
if the user want to run the K-i (\code{K_i = TRUE}) scan in parallel.
Default = NULL.}

\item{verbose}{\code{Logical} indicating if function outputs should be printed.
This argument is only used for the K-i model (\code{K_i = TRUE}).
Default = FALSE.}
}
\value{
Return:

\item{G_res}{Object of class \code{G_res} representing a data.frame with four
columns: marker identifier, chromosome, position in cM and -log10(p-value).}
}
\description{
Perform a genetic similarity matrix (GSM) genome wide association test using
a kinship matrix to correct for the genetic background.
}
\details{
The first step of the analysis is to partition the genome into a certain
number of markers sets. Several strategies are possibles. \code{hap}
determine the size of the block or window. It can be defined in number of
markers (\code{hap.unit = 1}) or in cM (\code{hap.unit = 2}). The blocks
can be adjacent (no overlapping, e.g. b1 = 123, b2 = 456, etc.) or can
be deterimed by sliding window (\code{sliding.window = TRUE}). When user
choose sliding window it is also possible to define the number of marker
between the two centers of a window using \code{gap}.

The genome scan is based on the sequence kernel association test (SKAT)
proposed by Wu et al. (2011) calling the SKAT package. This means that the
function determines the genetic relatedness between lines using a set of
markers and a kernel function. The variance covariance structure determined
by the kernel function is then used to test for association between the
phenotype and a tested region. The marker VCOV is assoiciated to a random
QTL term. The procedure is repeated at different positions along the genome
to perform a full genome scan.

Three models are possible. The first one does not correct for the genetic
background using a kinship matrix (\code{kin.correct = FALSE}). The second
and the third models correct for the genetic background using a
kinship matrix (\code{kin.correct = TRUE}) that can include all markers
(\code{K_i = FALSE}) or remove the markers on the currently scanned
chromsome (\code{K_i = TRUE}). All models include a fixed cross-specific
intercept term. The model using the kinship correction are fitted using
the EMMA algorithm proposed by Kang et al. (2008).

By default, the kinship is computed using the method of Astle and Balding
(2009). It is possible to compute a linkage disequilibrium adjusted kinship
(LDAK) kinship matrix using the method of Speed et al. (2012) by introducing
the weights computed with the function \code{\link{LDAK_weights}}. The model
can be fitted using the kinship containing all markers or removing
the markers of the scanned chromosome (\code{K_i = TRUE}).
}
\examples{

data("EUNAM_gp")
data("EUNAM_LD_weights")

# for gp object construction and weights computation see examples of
# LDAK_weights()

\dontrun{

  plink.dir <- "/home/.../PLINK"

  SSD_file <- generate_SSD_file(gp = EUNAM_gp, out.dir = getwd(), prefix = "Test",
                                plink.dir = plink.dir, hap = 1, hap.unit = 2,
                                sliding.window = TRUE)

  res <- GWAS_GSM(SSD_file = SSD_file, gp = EUNAM_gp, kin.correct = TRUE,
                  weights = EUNAM_LD_weights, K_i = TRUE)

  plot_manhattan(res)
  plot_qq(res)

}

}
\references{
Astle, W., & Balding, D. J. (2009). Population structure and cryptic
relatedness in genetic association studies. Statistical Science, 451-471.

Kang, H. M., Zaitlen, N. A., Wade, C. M., Kirby, A., Heckerman, D., Daly,
M. J., & Eskin, E. (2008). Efficient control of population structure in
model organism association mapping. Genetics, 178(3), 1709-1723.

Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.

Wu, M. C., Lee, S., Cai, T., Li, Y., Boehnke, M., & Lin, X. (2011).
Rare-variant association testing for sequencing data with the sequence kernel
association test. The American Journal of Human Genetics, 89(1), 82-93.
}
\author{
Vincent Garin
}
