% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_haplo.R
\name{GWAS_haplo}
\alias{GWAS_haplo}
\title{Haplotype Association Test}
\usage{
GWAS_haplo(haplo.block, haplo.term = "fixed", gp, map, trait,
  weights = NULL, power = -1, K_i = TRUE)
}
\arguments{
\item{haplo.block}{\code{List} containing the required data to compute an
haplotype model. This object can be obtained using the function
\code{\link{haplo_blocks}}.}

\item{haplo.term}{\code{Character} variable indicating if the haplotype term
should be computed as fixed ("fixed") or random ("random").
Default = "fixed".}

\item{gp}{\code{gpData} object with elements geno coded 0 1 2 and family
\strong{containing the list of markers that will be used to compute the
kinship matrix}}

\item{map}{Three columns \code{data.frame} corresponding to the list of
markers present in \code{gp}. The three columns represent the marker
identifier, the chromosome, and the marker position (cM or bp).}

\item{trait}{\code{Numerical} vector of phenotypic trait values.}

\item{weights}{Two columns \code{data.frame} containing the marker
identifiers and the weights marker weights. For example the weights obtained
with the LDAK program (see \code{\link{LDAK_weights}}). By default, the
kinship matrix is computed unweighted or all weights are equal to 1, which
correspond to the Astle and Balding kinship matrix}

\item{power}{\code{Numerical} value specifying the value of the
parameter for marker scores standardization. The column of the marker matrix
(X.j) are multiplied by var(X.j)^(power/2). It correspond to alpha in the
formula .Default = -1.}

\item{K_i}{\code{Logical} specifying if the kinship correction should be done
by removing the markers of the scanned chromosome. Default = TRUE.}
}
\value{
Return:

\item{res}{Data.frame with in row the haplotype blocks and in column the
LRT and Wald statistics of each haplotype blocks.}
}
\description{
Perform an haplotype genome wide association test using a kinship matrix
to correct for the genetic background.
}
\details{
The function is an adaptation of the Fixed-A and Random-A models proposed
by Wei and Xu (2016). We modified the code of the function \code{magicScan}
from the \code{magicQTL} package
(\url{http://www.genetics.org/content/202/2/471.supplemental})
(Wei and Xu, 2016). We replace the founder QTL term by an haplotype QTL term.
As in \code{magicQTL}, the QTL term can be fitted as fixed or random. The
model is fitted using the EMMA algorithm proposed by Kang et al. (2008).

A correction for the genetic background is applied using a kinship matrix.
By default, the kinship is computed using the method of Astle and Balding
(2009). It is possible to compute a linkage disequilibrium adjusted kinship
(LDAK) kinship matrix using the method of Speed et al. (2012) by introducing
the weights computed with the function \code{\link{LDAK_weights}}. The model
can be fitted using the kinship containing all markers or removing
the markers of the scanned chromosome (\code{K_i = TRUE}). In the sitation,
where the markers of the scanned chromosomes are removed (\code{K_i = TRUE}),
if the computation failed for one or several chromosomes, their results
are set to 0 and a warning message is produced at the end of the scan.
}
\references{
Astle, W., & Balding, D. J. (2009). Population structure and cryptic
relatedness in genetic association studies. Statistical Science, 451-471.

Kang, H. M., Zaitlen, N. A., Wade, C. M., Kirby, A., Heckerman, D., Daly,
M. J., & Eskin, E. (2008). Efficient control of population structure in
model organism association mapping. Genetics, 178(3), 1709-1723.

Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.

Wei, J., & Xu, S. (2016). A random-model approach to QTL mapping in
multiparent advanced generation intercross (MAGIC) populations.
Genetics, 202(2), 471-486.
}
\author{
Vincent Garin, Julong Wei and Shizhong Xu
(original underlying functions)
}
