% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_SNP.R
\name{GWAS_SNP}
\alias{GWAS_SNP}
\title{Single SNP GWAS}
\usage{
GWAS_SNP(gp, trait = 1, weights = NULL, power = -1, mk.sel = NULL,
  K_i = TRUE, n.cores = NULL, verbose = FALSE)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2, map with
marker position in cM, phenotype and family indicator.}

\item{trait}{\code{Numerical} or \code{character} indicator to specify which
trait of the gp object should be used. Default = 1.}

\item{weights}{object of class \code{LD_wgh} obtained by the function
LDAK_weights() representing a data.frame with two columns: the marker
identifier and the LD adjusted weights. These weight will be used to compute
a LDAK as defined by Speed et al. (2012) for the genetic background
adjustement. Default = NULL.}

\item{power}{\code{Numerical} value specifying the value of the
parameter for marker scores standardization. The column of the marker matrix
(X.j) are multiplied by var(X.j)^(power/2). It correspond to alpha in the
formula. Default = -1.}

\item{mk.sel}{\code{Character vector} specifying a list of marker to use
for the kinship matrix computation. By default, the function use all markers
of the \code{gp}.}

\item{K_i}{\code{Logical} specifying if the kinship correction should be done
by removing the markers of the scanned chromosome. Default = TRUE.}

\item{n.cores}{\code{Numeric} value indicating the number of core to be used
if the user want to run the K-i (\code{K_i = TRUE}) scan in parallel.
Default = NULL.}

\item{verbose}{\code{Logical} indicating if function outputs should be printed.
Default = FALSE.}
}
\value{
Return:

\item{G.res}{Object of class \code{G_res} representing a data.frame with four
columns: marker identifier, chromosome, position in cM and -log10(p-value).}
}
\description{
Perform a single SNP genome wide association test using a kinship matrix
to correct for the genetic background.
}
\details{
The function is a wrapper for the \code{GWAS} function from the \code{sommer}
package (Covarrubias-Pazaran, 2016). The model is fitted using the EMMA
algorithm proposed by Kang et al. (2008).

By default, the kinship matrix is computed using the method of Astle and
Balding (2009). It is possible to compute a linkage disequilibrium adjusted
kinship (LDAK) kinship matrix using the method of Speed et al. (2012) by
introducing the weights computed with the function \code{\link{LDAK_weights}}.
The model can be fitted using the kinship containing all markers or removing
the markers of the scanned chromosome (\code{K_i = TRUE}).
}
\references{
Astle, W., & Balding, D. J. (2009). Population structure and cryptic
relatedness in genetic association studies. Statistical Science, 451-471.

Covarrubias-Pazaran G. 2016. Genome assisted prediction of quantitative traits
using the R package sommer. PLoSONE 11(6):1-15.

Kang, H. M., Zaitlen, N. A., Wade, C. M., Kirby, A., Heckerman, D., Daly,
M. J., & Eskin, E. (2008). Efficient control of population structure in model
organism association mapping. Genetics, 178(3), 1709-1723.

Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.
}
\author{
Vincent Garin
}
