% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAK_kinship.R
\name{LDAK_kinship}
\alias{LDAK_kinship}
\title{Compute linkage desequilibrium adjusted kinship}
\usage{
LDAK_kinship(ldak.dir, weights.loc, bed.file.loc, out.dir = getwd(),
  power = -1, K_i = NULL, map = NULL)
}
\arguments{
\item{ldak.dir}{directory where the ldak executable program is located}

\item{weights.loc}{path to the LDAK weights files (output of function
\code{\link{LDAK_weights}}). For example, /home/.../weights.}

\item{bed.file.loc}{path to the .bed files (output of function
\code{\link{write_plink_bed}}) without .bed extension. For example,
/home/.../my_file.}

\item{out.dir}{output directory where temporary file will be saved.
These files will be removed. Default = getwd().}

\item{power}{\code{Numerical} value specifying the value of the
parameter for marker scores standardization. The column of the marker matrix
(X.j) are multiplied by var(X.j)^(power/2) .Default = -1.}

\item{K_i}{\code{Numerical} value specifying a unique chromosome number that
should be removed from the kinship computation. By default \code{K_i = NULL},
which means that the kinship is computed using all markers.}

\item{map}{If \code{K_i} is not NULL, \code{data.frame} map information with
at least a colum for marker identifier labeled \code{'mk.id'}, and one column
for chromosome indicator labeled \code{'chr'}. \strong{The marker identificer
must be the same as the one of the weight file.}. Default = NULL.}
}
\value{
Return:

\item{K}{kinship matrix computed with the LDAK weights.}
}
\description{
Compute linkage disequilibrium adjusted kinship using the program LDAK.
(http://dougspeed.com/ldak/). The function is a wrapper for the LDAK
software.
}
\references{
Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.
}
\author{
Vincent Garin
}
