% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAK_kinship2.R
\name{LDAK_kinship2}
\alias{LDAK_kinship2}
\title{Compute linkage desequilibrium adjusted kinship}
\usage{
LDAK_kinship2(gp, weights, power = -1, K_i = NULL)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2.}

\item{weights}{Four columns \code{data.frame} containing LDAK weights
information. The columns must be: 1) marker identifier, chromosome,
position (in cM or bp), LDAK weights (see \code{\link{LDAK_weights}}).}

\item{power}{\code{Numerical} value specifying the value of the
parameter for marker scores standardization. The column of the marker matrix
(X.j) are multiplied by var(X.j)^(power/2) .Default = -1.}

\item{K_i}{\code{Numerical} value specifying a unique chromosome number that
should be removed from the kinship computation. By default \code{K_i = NULL},
which means that the kinship is computed using all markers.}
}
\value{
Return:

\item{K}{Kinship matrix computed with the LDAK weights.}
}
\description{
Compute linkage disequilibrium adjusted kinship using the weights computed by
the program LDAK. (http://dougspeed.com/ldak/).
}
\details{
The results return by the function are two times the results given by LDAK
using --calc-kins-direct.
}
\references{
Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.
}
\author{
Vincent Garin
}
