% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAK_weights.R
\name{LDAK_weights}
\alias{LDAK_weights}
\title{Compute Linkage disequilibrium markers weights}
\usage{
LDAK_weights(gp, bp.pos, out.dir, plink.dir, ldak.dir, verbose = FALSE)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2, map with
marker position in cM, phenotype and family indicator.}

\item{bp.pos}{\code{Numeric vector} of marker physical base pair positions.}

\item{out.dir}{output directory where the ldak weights will be stored.
Defaut = working directory.}

\item{plink.dir}{directory where the plink executable program is located.}

\item{ldak.dir}{directory where the ldak executable program is located}

\item{verbose}{\code{Logical} indicating if function outputs should be printed.
Default = FALSE}
}
\value{
Return:

An object of class \code{LD_wgh} which is a \code{data.frame} with two columns
containing: the marker identifier and the computed LDAK weight.
}
\description{
Compute linkage disequilibrium weights using the program LDAK.
The function is a wrapper for the LDAK software.
}
\details{
This function use the LDAK 5 beta executable file that can be download here:
http://dougspeed.com/ldak/ and the plink executable file that can be downloaded
here: http://zzz.bwh.harvard.edu/plink/download.shtml.
}
\examples{

# form the gpData object

library(synbreed)

data("EUNAM_geno")
data("EUNAM_map")
data("EUNAM_pheno")

cross.ind <- substr(rownames(EUNAM_pheno), 1, 5)

map_gp <- EUNAM_map[, 2:3]
colnames(map_gp) <- c("chr", "pos")
rownames(map_gp) <- EUNAM_map[, 1]

family <- data.frame(cross.ind)
rownames(family) <- rownames(EUNAM_pheno)

gp <- create.gpData(pheno = EUNAM_pheno, geno = EUNAM_geno, family = family,
                    map = map_gp)

# compute the LDAK weights

\dontrun{

  plink.dir = "/home/.../PLINK"
  ldak.dir <- "/home/.../LDAK"

  wgt <- LDAK_weights(gp = gp, bp.pos = EUNAM_map[, 4], out.dir = getwd(),
                      plink.dir = plink.dir, ldak.dir = ldak.dir)

}

}
\references{
Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.
}
\author{
Vincent Garin
}
