% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpp_kinship.R
\name{mpp_kinship}
\alias{mpp_kinship}
\title{Compute linkage desequilibrium adjusted kinship}
\usage{
mpp_kinship(gp, weights = NULL, power = -1, mk.sel = NULL)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2.}

\item{weights}{Two columns \code{data.frame} containing the marker
identifiers and the weights marker weights. For example the weights obtained
with the LDAK program (see \code{\link{LDAK_weights}}). By default, the
kinship matrix is computed unweighted or all weights are equal to 1, which
correspond to the Astle and Balding kinship matrix}

\item{power}{\code{Numerical} value specifying the value of the
parameter for marker scores standardization. The column of the marker matrix
(X.j) are multiplied by var(X.j)^(power/2). It correspond to alpha in the
formula .Default = -1.}

\item{mk.sel}{\code{Character vector} specifying a list of marker to use
for the kinship matrix computation. By default, the function use all markers
of the \code{gpData} object.}
}
\value{
Return:

\item{K}{Kinship matrix computed with the LDAK weights.}
}
\description{
Compute linkage disequilibrium adjusted kinship using the weights computed by
the program LDAK. (http://dougspeed.com/ldak/).
}
\details{
The formula used to compute the kinship is the following:

(1/sum(wj)) [sqrt(wj)(Xj - 2pj)][sqrt(wj)(Xj - 2pj)]'/[(2pj(1-pj))^alpha]

The results return by the function are two times the results given by LDAK
using --calc-kins-direct.
}
\references{
Astle, W., & Balding, D. J. (2009). Population structure and cryptic
relatedness in genetic association studies. Statistical Science, 451-471.

Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.
}
\author{
Vincent Garin
}
