% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.G_res.R
\name{plot.G_res}
\alias{plot.G_res}
\title{Plot Association test results}
\usage{
\method{plot}{G_res}(x, pl.type = "man", ...)
}
\arguments{
\item{x}{Object of class \code{G_res} obtained with function
\code{\link{GWAS_SNP}}, \code{\link{GWAS_haplo}}, \code{\link{GWAS_GSM}}.}

\item{pl.type}{\code{Character} expression. Either "man" for a manhattan plot
or "qq" for a QQ plot. Default = "man".}

\item{...}{further arguments passed to or from other methods.}
}
\description{
S3 \code{plot} method for object of class \code{G_res}. The function
return either a manhattan plot (\code{type = "man"}) or a QQ plot
(\code{type = "qq"})
}
\details{
The function call functions from the package qqman (Turner, 2014).
}
\references{
Turner, S.D. qqman: an R package for visualizing GWAS results using Q-Q and
manhattan plots. biorXiv DOI: 10.1101/005165 (2014).
}
\author{
Vincent Garin
}
