% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot of Association Test results}
\usage{
plot_manhattan(res, ...)
}
\arguments{
\item{res}{Object of class \code{G_res} obtained with function
\code{\link{GWAS_SNP}}, \code{\link{GWAS_haplo}}, \code{\link{GWAS_GSM}}.}

\item{...}{Arguments passed on to other plot/points functions}
}
\description{
Plot the results of an association test using manhattan plot using the
\code{manhattan} function of the qqman package.
}
\author{
Vincent Garin
}
