% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_AssTest.R
\name{SNP_AssTest}
\alias{SNP_AssTest}
\title{Single SNP Association Test}
\usage{
SNP_AssTest(gp, trait, map, weights = NULL, power = -1, mk.sel = NULL,
  K_i = FALSE, n.cores = 1)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2 and family.}

\item{trait}{\code{Numerical} vector of phenotypic trait values.}

\item{map}{If Three columns \code{data.frame} with marker identifier,
chromosome, and marker position (cM or bp).}

\item{weights}{Two columns \code{data.frame} containing the marker
identifiers and the weights marker weights. For example the weights obtained
with the LDAK program (see \code{\link{LDAK_weights}}). By default, the
kinship matrix is computed unweighted or all weights are equal to 1, which
correspond to the Astle and Balding kinship matrix}

\item{power}{\code{Numerical} value specifying the value of the
parameter for marker scores standardization. The column of the marker matrix
(X.j) are multiplied by var(X.j)^(power/2). It correspond to alpha in the
formula .Default = -1.}

\item{mk.sel}{\code{Character vector} specifying a list of marker to use
for the kinship matrix computation. By default, the function use all markers
of the \code{gpData} object.}

\item{K_i}{\code{Logical} specifying if the kinship correction should be done
by removing the markers of the scanned chromosome. Default = FALSE.}

\item{n.cores}{Default = 1.}
}
\value{
Return:

\item{results}{The results}
}
\description{
Perform a single SNP genome wide association test using a kinship matrix
to correct for the genetic background. By default, the kinship is computed
using the method of Astle and Balding (2009). It is possible to compute
a linkage disequilibrium adjusted kinship (LDAK) kinship matrix using the
method of Speed et al. (2012) by introducing the weights computed with the
function \code{\link{LDAK_weights}}.
}
\references{
Astle, W., & Balding, D. J. (2009). Population structure and cryptic
relatedness in genetic association studies. Statistical Science, 451-471.

Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.
}
\author{
Vincent Garin
}
