% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plink_bed.R
\name{write_plink_bed}
\alias{write_plink_bed}
\title{Produce plink bed data}
\usage{
write_plink_bed(gp, map, out.dir, prefix, plink.dir)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2 and family}

\item{map}{Four columns \code{data.frame} with chromosome, marker id,
getic position in cM and physical position in bp.}

\item{out.dir}{output directory where the .bed files will be saved.}

\item{prefix}{\code{Character}. Prefix for the plink input file}

\item{plink.dir}{directory where the plink executable program is located.}
}
\value{
Return:

the function save the plink .bed files (.bed, .bim and .fam) in the directory
specified in argument \code{out.dir}
}
\description{
Produce plink bed data
}
\author{
Vincent Garin
}
