% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_haplo.R
\name{GWAS_haplo}
\alias{GWAS_haplo}
\title{Haplotype Association Test}
\usage{
GWAS_haplo(haplo.block, haplo.term = "fixed", gp, trait = 1,
  weights = NULL, power = -1, K_i = TRUE, n.cores = NULL,
  verbose = FALSE)
}
\arguments{
\item{haplo.block}{Object of class \code{hap_bl} obtained with the function
\code{\link{haplo_blocks}} containing the haplotype matrices necessary to
compute the haplotype model.}

\item{haplo.term}{\code{Character} variable indicating if the haplotype term
should be computed as fixed ("fixed") or random ("random").
Default = "fixed".}

\item{gp}{\code{gpData} object with elements geno coded 0 1 2, map with
marker position in cM, phenotype and family. \strong{containing the list of
markers that will be used to compute the kinship matrix}.}

\item{trait}{\code{Numerical} or \code{character} indicator to specify which
trait of the gp object should be used. Default = 1.}

\item{weights}{object of class \code{LD_wgh} obtained by the function
LDAK_weights() representing a data.frame with two columns: the marker
identifier and the LD adjusted weights. These weight will be used to compute
a LDAK as defined by Speed et al. (2012) for the genetic background
adjustement. Default = NULL. By default, the kinship matrix is computed
unweighted or all weights are equal to 1, which correspond to the Astle and
Balding kinship matrix.}

\item{power}{\code{Numerical} value specifying the value of the
parameter for marker scores standardization. The column of the marker matrix
(X.j) are multiplied by var(X.j)^(power/2). It correspond to alpha in the
formula .Default = -1.}

\item{K_i}{\code{Logical} specifying if the kinship correction should be done
by removing the markers of the scanned chromosome. Default = TRUE.}

\item{n.cores}{\code{Numeric} value indicating the number of core to be used
if the user want to run the K-i (\code{K_i = TRUE}) scan in parallel.
Default = NULL.}

\item{verbose}{\code{Logical} indicating if function outputs should be printed.
Default = FALSE.}
}
\value{
Return:

\item{G.res}{Object of class \code{G_res} representing a data.frame with four
columns: marker identifier, chromosome, position in cM and -log10(p-value).}
}
\description{
Perform an haplotype genome wide association test using a kinship matrix
to correct for the genetic background.
}
\details{
The function is an adaptation of the Fixed-A and Random-A models proposed
by Wei and Xu (2016). We modified the code of the function \code{magicScan}
from the \code{magicQTL} package
(\url{http://www.genetics.org/content/202/2/471.supplemental})
(Wei and Xu, 2016). We replace the founder QTL term by an haplotype QTL term.
As in \code{magicQTL}, the QTL term can be fitted as fixed or random. The
model is fitted using the EMMA algorithm proposed by Kang et al. (2008).

A correction for the genetic background is applied using a kinship matrix.
By default, the kinship is computed using the method of Astle and Balding
(2009). It is possible to compute a linkage disequilibrium adjusted kinship
(LDAK) kinship matrix using the method of Speed et al. (2012) by introducing
the weights computed with the function \code{\link{LDAK_weights}}. The model
can be fitted using the kinship containing all markers or removing
the markers of the scanned chromosome (\code{K_i = TRUE}). In the sitation,
where the markers of the scanned chromosomes are removed (\code{K_i = TRUE}),
if the computation failed for one or several chromosomes, their results
are set to 0 and a warning message is produced at the end of the scan.
}
\references{
Astle, W., & Balding, D. J. (2009). Population structure and cryptic
relatedness in genetic association studies. Statistical Science, 451-471.

Kang, H. M., Zaitlen, N. A., Wade, C. M., Kirby, A., Heckerman, D., Daly,
M. J., & Eskin, E. (2008). Efficient control of population structure in
model organism association mapping. Genetics, 178(3), 1709-1723.

Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.

Wei, J., & Xu, S. (2016). A random-model approach to QTL mapping in
multiparent advanced generation intercross (MAGIC) populations.
Genetics, 202(2), 471-486.
}
\author{
Vincent Garin, Julong Wei and Shizhong Xu
(original underlying functions)
}
