% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_blocks.R
\name{haplo_blocks}
\alias{haplo_blocks}
\title{Determine haplotype blocks and haplotypes}
\usage{
haplo_blocks(gp, map, out.dir, hap = 1, hap.unit = 1,
  sliding.window = FALSE, LD.based = FALSE, ld.threshold = 0.9)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2.}

\item{map}{Four columns \code{data.frame} with marker id, chromosome,
getic position in cM and physical position in bp.}

\item{out.dir}{output directory where temporary files will be stored.}

\item{hap}{Defines either a number of markers (if \code{hap.unit=1}) or a
genetic distance in cM (if \code{hap.unit=2}) that should be used for building
haplo blocks. Default = 1.}

\item{hap.unit}{Set to 1 for building haplo blocks based on number of
markers, set to 2 for building haplo blocks based on markers lying in a
specified stretch of the chromosome (in cM). Default = 1.}

\item{sliding.window}{NOT AVAILABLE NOW. \code{logical} value specifying if
the haplotype blocks should be construct using a sliding window. ... .
Default = FALSE.

[...]}

\item{LD.based}{NOT AVAILABLE NOW. \code{logical} value specifying if the
haplotype blocks should be constructed based on LD. Default = FALSE.}

\item{ld.threshold}{NOT AVAILABLE NOW.}
}
\value{
Return a list containing the following elements:

\item{haplo.geno}{Data frame with in row the haplotype block and in column
the genotypes. For each haplotype, the haplotype alleles are coded with
a different haplotype score.}

\item{haplo.map}{Four column map data frame with the haplotpype block id
the chromsome, the cM and the bp position of the haplotype block.}
}
\description{
Determine haplotype block and haplotypes using the SelectionTools package
(\url{http://www.population-genetics.de/~frisch-m/}).
The haplotype blocks can be determined based on several criteria:
number of markers, distance in cM or LD based.
}
\author{
Vincent Garin
}
