% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAK_weights.R
\name{LDAK_weights}
\alias{LDAK_weights}
\title{Compute Linkage disequilibrium markers weights}
\usage{
LDAK_weights(gp, map, out.dir, prefix, plink.dir, ldak.dir, kinship = FALSE)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2 and family}

\item{map}{Four columns \code{data.frame} with chromosome, marker id,
getic position in cM and physical position in bp. with column names being
"mk.id", "chr", "cM", "bp".}

\item{out.dir}{output directory where the ldak weights will be stored}

\item{prefix}{\code{Character}. Prefix for the plink input file}

\item{plink.dir}{directory where the plink executable program is located.}

\item{ldak.dir}{directory where the ldak executable program is located}

\item{kinship}{\code{Logical} value specifying if the kinshi matrix should be
returned. Default = FALSE.}
}
\value{
Return:

the function save the LDAK weights in the directory specified in argument
\code{out.dir}.

If \code{kinship = TRUE}, kinship matrix computed with the LDAK weights.
}
\description{
Compute linkage disequilibrium weights using the program LDAK.
(http://dougspeed.com/ldak/). The function is a wrapper for the LDAK
software.
}
\references{
Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.
}
\author{
Vincent Garin
}
