% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAK_weights.R
\name{LDAK_weights}
\alias{LDAK_weights}
\title{Compute Linkage disequilibrium markers weights}
\usage{
LDAK_weights(gp, bp.pos, out.dir, plink.dir, ldak.dir, verbose = FALSE)
}
\arguments{
\item{gp}{\code{gpData} object with elements geno coded 0 1 2, family and
map with position in cM.}

\item{bp.pos}{\code{Numeric vector} of marker physical base pair positions.}

\item{out.dir}{output directory where the ldak weights will be stored.
Defaut = working directory.}

\item{plink.dir}{directory where the plink executable program is located.}

\item{ldak.dir}{directory where the ldak executable program is located}

\item{verbose}{\code{Logical} indicating if function outputs should be printed.}
}
\value{
Return:

An object of class \code{LD_wgh} which is a \code{data.frame} with two columns
containing: the marker identifier and the computed LDAK weight.
}
\description{
Compute linkage disequilibrium weights using the program LDAK.
The function is a wrapper for the LDAK software.
}
\details{
This function use the LDAK 5 beta executable file that can be download here:
http://dougspeed.com/ldak/ and the plink executable file that can be downloaded
here: http://zzz.bwh.harvard.edu/plink/download.shtml.
}
\references{
Speed, D., Hemani, G., Johnson, M. R., & Balding, D. J. (2012).
Improved heritability estimation from genome-wide SNPs. The American Journal
of Human Genetics, 91(6), 1011-1021.
}
\author{
Vincent Garin
}
