% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot of Association Test results}
\usage{
plot_manhattan(res)
}
\arguments{
\item{res}{result of function \code{SNP_AssTest}.}
}
\description{
Plot the results of an association test using manhattan plot using the
\code{manhattan} function of the qqman package.
}
\author{
Vincent Garin
}
