% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{QQ plot of Association Test results}
\usage{
plot_qq(res)
}
\arguments{
\item{res}{result of function \code{SNP_AssTest}.}
}
\description{
Plot the results of an association test using manhattan plot using the
\code{manhattan} function of the qqman package.
}
\author{
Vincent Garin
}
